fix: adding dependencies
I wanted to start working on adding a feature, however without making
changes to the ts.config.js, eslintrc, and add dependencies the
build process would not complete. Modules would not resolve for two
reasons: The first was that there did not seem to be a correct setup
for working with the packages folder out of the box; the second was
that I expect the developer has installed many packages globally on
the dev machine. This however stops it from working on others.

The tests still fail due to expecting Redis server among others, however
the build does pass.